+++ CHANGELOG.md
@@ -41,6 +41,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Nomenclature match updates [#2176]
 - Navigation key combination for radial annotator [#3233]
 - Truncate smart selector lists
+- Allow compare n objects in collection object match [#3238]
 
 [#3190]: https://github.com/SpeciesFileGroup/taxonworks/issues/3190
 [#2176]: https://github.com/SpeciesFileGroup/taxonworks/issues/2176
@@ -53,6 +54,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 [#3192]: https://github.com/SpeciesFileGroup/taxonworks/issues/3192
 [#3203]: https://github.com/SpeciesFileGroup/taxonworks/issues/3203
 [#3212]: https://github.com/SpeciesFileGroup/taxonworks/issues/3212
+[#3238]: https://github.com/SpeciesFileGroup/taxonworks/issues/3238
 
 ## [0.29.6] - 2022-11-08